+++ changelog.md
@@ -1,10 +1,11 @@
 # Changelog
 
 ## Upcoming
-* Healing potions now heal 3*2^level instead of 2*2^level per level (like pre-1.6.1)
+* Healing potions now heal `3*2^level` instead of `2*2^level` per level (like pre-1.6.1)
 * Slightly increased spawners delay when empowered and increased experience from broken spawners
 * Fixed Plant growth modifiers not working properly
   * Fixes nether wart not growing and some plants growing slower than they should
+* Fixed infinite arrows with Infinity and Allurement
 
 ## 3.10.0
 * Added Echo Lantern